improving ai connector flow
