+++ x-pack/plugins/security_solution/public/onboarding/components/onboarding_body/cards/assistant/translations.ts
@@ -35,3 +35,10 @@ export const ASSISTANT_CARD_CALLOUT_INTEGRATIONS_BUTTON = i18n.translate(
     defaultMessage: 'Add integrations step',
   }
 );
+
+export const ASSISTANT_CARD_CREATE_NEW_CONNECTOR_POPOVER = i18n.translate(
+  'xpack.securitySolution.onboarding.assistantCard.calloutIntegrationsButton',
+  {
+    defaultMessage: 'Create new connector',
+  }
+);
